pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "guin"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.26",
    "fastmcp>=3.1",
    "anthropic>=0.45",
    "langgraph>=0.3.5,<0.4",
    "langgraph-codeact>=0.1.3",
    "langchain>=0.3",
    "langchain-text-splitters>=0.3",
    "nipype>=1.8",
    "pybids>=0.16",
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    "click>=8.1",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "websockets>=13",
    "pydantic>=2.7",
    "httpx>=0.27",
    "prov>=2.0",
    "rdflib>=7.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "rich>=13",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.6", "mypy>=1.11", "pre-commit>=3"]

[project.scripts]
guin = "guin.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/guin"]

[tool.mypy]
python_version = "3.11"
explicit_package_bases = true
mypy_path = "src"
packages = ["guin"]
